db: make perm columns a Perm instead of uint8

+++ main.go
@@ -41,7 +41,7 @@ func main() {
 	//
 	// create server 'Owner' Role: uneditable, and has all perms always
 
-	pa := uint8(db.PermAllow)
+	pa := db.PermAllow
 	db.BuiltInRoles["o"] = &db.Role{
 		0, "o", 0, "Owner", "", pa, pa, false, pa,
 	}
